Making files for each size of tape
