+++ src/utils/Bootstrap.cpp
@@ -29,7 +29,9 @@ int Bootstrap::init(int argc, char** argv)
     Parser parser(args_handler.get_string_file());
 
     io_lib = std::string(getenv("HOME")) + "/.brain/lib/";
-    io_lib += ArgsOptions::instance()->get_io_option() == IO_REPL ? "io_repl.ll" : "io.ll"; 
+    io_lib += ArgsOptions::instance()->get_io_option() == IO_REPL ?
+        "io_repl.ll" :
+        "io_" + std::to_string(ArgsOptions::instance()->get_cell_bitsize()) + ".ll";
 
     if (ArgsOptions::instance()->has_option(BO_IS_EMITTING_CODE)) {
         parser.ast_code_gen();
